fetch via Promise, moved to extra file
