+++ package.json
@@ -43,6 +43,7 @@
     "d3-transition": "^1.2.0",
     "leaflet": "^1.4.0",
     "leaflet-hash": "^0.2.1",
+    "lodash": "^4.17.11",
     "whatwg-fetch": "^3.0.0"
   }
 }
