+++ src/js/index.js
@@ -13,7 +13,7 @@ import {timeMinute} from 'd3-time';
 import {interval, timeout} from 'd3-timer';
 import {timeFormatLocale, timeParse} from 'd3-time-format';
 import {median} from 'd3-array';
-import 'whatwg-fetch';
+import api from './feinstaub-api';
 
 const d3 = Object.assign({}, d3_Selection, d3_Hexbin);
 
@@ -90,27 +90,6 @@ const panelIDs = {
     "Temperature": [4, 3],
     "Humidity": [6, 5],
     "Pressure": [8, 7]
-};
-
-const pm_sensors = {
-	"SDS011": true,
-	"SDS021": true,
-	"PMS1003": true,
-	"PMS3003": true,
-	"PMS5003": true,
-	"PMS6003": true,
-	"PMS7003": true,
-	"HPM": true,
-	"SPS30": true,
-};
-
-const thp_sensors = {
-	"DHT22": true,
-	"BMP180": true,
-	"BMP280": true,
-	"BME280": true,
-	"HTU21B": true,
-	"DS18B20": true,
 };
 
 const div = d3.select("#sidebar").append("div").attr("id", "table").style("display", "none");
@@ -431,13 +410,19 @@ The values are refreshed every 5 minutes in order to fit with the measurement fr
 
 //	REVOIR ORDRE DANS FONCTION READY
     function retrieveData() {
-        fetchURL("https://maps.luftdaten.info/data/v2/data.dust.min.json").then(function (json) {
-            ready(json, 1);
-            fetchURL("https://maps.luftdaten.info/data/v2/data.24h.json").then(function (json) {
-                ready(json, 2)
+		api.getAllSensors("https://maps.luftdaten.info/data/v2/data.dust.min.json",1).then(function(result) {
+			hmhexaPM_aktuell = result.cells;
+			if (result.timestamp > timestamp_data) timestamp_data = result.timestamp;
+			ready(1);
+			api.getAllSensors("https://maps.luftdaten.info/data/v2/data.24h.json",2).then(function (result) {
+				hmhexaPM_AQI = result.cells;
+				if (result.timestamp > timestamp_data) timestamp_data = result.timestamp;
+                ready(2)
             });
-            fetchURL("https://maps.luftdaten.info/data/v2/data.temp.min.json").then(function (json) {
-                ready(json, 3)
+            api.getAllSensors("https://maps.luftdaten.info/data/v2/data.temp.min.json",3).then(function (result) {
+				hmhexa_t_h_p = result.cells;
+				if (result.timestamp > timestamp_data) timestamp_data = result.timestamp;
+                ready(3)
             });
         });
         console.log('retrieving data')
@@ -480,92 +465,8 @@ function switchLegend(val) {
     d3.select('#legend_' + val).style("display", "block");
 }
 
-function checkValues(obj,sel) {
-    let result = false;
-    if (obj !== undefined && typeof (obj) === 'number' && !isNaN(obj)) {
-        if ((sel === "Humidity") && (obj >= 0) && (obj <= 100)) {
-            result = true;
-        } else if ((sel === "Temperature") && (obj <= 70 && obj >= -50)) {
-            result = true;
-        } else if ((sel === "Pressure") && (obj >= 850) && (obj < 1200)) {
-            result = true;
-        } else if ((sel === "PM10") && (obj < 1900)) {
-            result = true;
-        } else if ((sel === "PM25") && (obj < 900)) {
-            result = true;
-        } else if (sel === "Official_AQI_US") {
-			result = true;
-		}
-    }
-    return result;
-}
-
-function ready(data, num) {
+function ready(num) {
     let timestamp;
-    if (num === 1) {
-        hmhexaPM_aktuell = data
-			.filter(function (item) {
-				return (typeof pm_sensors[item.sensor.sensor_type.name] != "undefined" && pm_sensors[item.sensor.sensor_type.name] && checkValues(parseInt(getRightValue(item.sensordatavalues, "P1")),"PM10") && checkValues(parseInt(getRightValue(item.sensordatavalues, "P2")),"PM25"));
-			})
-			.map(function (item) {
-				if (item.timestamp > timestamp_data) timestamp_data = item.timestamp;				
-				return {
-					"data": {
-						"PM10": parseInt(getRightValue(item.sensordatavalues, "P1")),
-						"PM25": parseInt(getRightValue(item.sensordatavalues, "P2"))
-					},
-					"id": item.sensor.id,
-					"latitude": item.location.latitude,
-					"longitude": item.location.longitude
-				}
-			})			
-    } else if (num === 2) {
-        hmhexaPM_AQI = data
-			.filter(function (item){
-				return (typeof pm_sensors[item.sensor.sensor_type.name] != "undefined" && pm_sensors[item.sensor.sensor_type.name]);
-			})
-			.map(function(item){
-	            if (item.timestamp > timestamp_data) timestamp_data = item.timestamp;
-                const data_in = {
-                    "PM10": parseInt(getRightValue(item.sensordatavalues, "P1")),
-                    "PM25": parseInt(getRightValue(item.sensordatavalues, "P2"))
-                };
-                const data_out = officialAQIus(data_in);
-                return {
-                    "data": {
-                        "Official_AQI_US": data_out.AQI,
-                        "origin": data_out.origin,
-                        "PM10_24h": data_in.PM10,
-                        "PM25_24h": data_in.PM25
-                    },
-                    "id": item.sensor.id,
-                    "latitude": item.location.latitude,
-                    "longitude": item.location.longitude
-				}
-			})
-			.filter(function (item){
-				return (checkValues(item.data.Official_AQI_US,"Official_AQI_US"));
-			})
-    } else {
-        // REVOIR LES TYPES DE SENSORS
-        hmhexa_t_h_p = data
-			.filter(function(item){
-				return (typeof thp_sensors[item.sensor.sensor_type.name] != "undefined" && thp_sensors[item.sensor.sensor_type.name]);
-			})
-			.map(function(item) {
-	            if (item.timestamp > timestamp_data) timestamp_data = item.timestamp;
-				return {
-					"data": {
-						"Pressure": parseInt(getRightValue(item.sensordatavalues, "pressure_at_sealevel")) / 100,
-						"Humidity": parseInt(getRightValue(item.sensordatavalues, "humidity")),
-						"Temperature": parseInt(getRightValue(item.sensordatavalues, "temperature"))
-					},
-					"id": item.sensor.id,
-					"latitude": item.location.latitude,
-					"longitude": item.location.longitude
-				}
-			})
-    }
 
     const dateParser = timeParse("%Y-%m-%d %H:%M:%S");
     timestamp = dateParser(timestamp_data);
@@ -587,11 +488,10 @@ function ready(data, num) {
     if (num === 3 && (user_selected_value === "Temperature" || user_selected_value === "Humidity" || user_selected_value === "Pressure")) {
         hexagonheatmap.initialize(scale_options[user_selected_value]);
         hexagonheatmap.data(hmhexa_t_h_p.filter(function (value) {
-            return checkValues(value.data[user_selected_value],user_selected_value);
+            return api.checkValues(value.data[user_selected_value],user_selected_value);
         }));
     }
     d3.select("#loading").style("display", "none");
-
 }
 
 function reload(val) {
@@ -608,19 +508,9 @@ function reload(val) {
         hexagonheatmap.data(hmhexaPM_AQI);
     } else if (val === "Temperature" || val === "Humidity" || val === "Pressure") {
         hexagonheatmap.data(hmhexa_t_h_p.filter(function (value) {
-            return checkValues(value.data[user_selected_value],user_selected_value);
+            return api.checkValues(value.data[user_selected_value],user_selected_value);
         }));
     }
-}
-
-function getRightValue(array, type) {
-    let value;
-    array.forEach(function (item) {
-        if (item.value_type === type) {
-            value = item.value;
-        }
-    });
-    return value;
 }
 
 //MENU
@@ -703,56 +593,6 @@ function sensorNr(data) {
     });
 }
 
-function aqius(val, type) {
-    let index;
-
-    if (val >= 0) {
-        if (type === 'PM10') {
-            if (parseInt(val) <= 54) {
-                index = calculate_aqi_us(50, 0, 54, 0, parseInt(val))
-            } else if (parseInt(val) <= 154) {
-                index = calculate_aqi_us(100, 51, 154, 55, parseInt(val))
-            } else if (parseInt(val) <= 254) {
-                index = calculate_aqi_us(150, 101, 254, 155, parseInt(val))
-            } else if (parseInt(val) <= 354) {
-                index = calculate_aqi_us(200, 151, 354, 255, parseInt(val))
-            } else if (parseInt(val) <= 424) {
-                index = calculate_aqi_us(300, 201, 424, 355, parseInt(val))
-            } else if (parseInt(val) <= 504) {
-                index = calculate_aqi_us(400, 301, 504, 425, parseInt(val))
-            } else if (parseInt(val) <= 604) {
-                index = calculate_aqi_us(500, 401, 604, 505, parseInt(val))
-            } else {
-                index = 500
-            }
-        }
-        if (type === 'PM25') {
-            if (val.toFixed(1) <= 12) {
-                index = calculate_aqi_us(50, 0, 12, 0, val.toFixed(1))
-            } else if (val.toFixed(1) <= 35.4) {
-                index = calculate_aqi_us(100, 51, 35.4, 12.1, val.toFixed(1))
-            } else if (val.toFixed(1) <= 55.4) {
-                index = calculate_aqi_us(150, 101, 55.4, 35.5, val.toFixed(1))
-            } else if (val.toFixed(1) <= 150.4) {
-                index = calculate_aqi_us(200, 151, 150.4, 55.5, val.toFixed(1))
-            } else if (val.toFixed(1) <= 250.4) {
-                index = calculate_aqi_us(300, 201, 250.4, 150.5, val.toFixed(1))
-            } else if (val.toFixed(1) <= 350.4) {
-                index = calculate_aqi_us(400, 301, 350.4, 250.5, val.toFixed(1))
-            } else if (val.toFixed(1) <= 550.4) {
-                index = calculate_aqi_us(500, 401, 550.4, 350.5, val.toFixed(1))
-            } else {
-                index = 500
-            }
-        }
-    }
-    return index;
-}
-
-function calculate_aqi_us(Ih, Il, Ch, Cl, C) {
-    return parseInt((((Ih - Il) / (Ch - Cl)) * (C - Cl)) + Il);
-}
-
 function displayGraph(id) {
 
     let inner_pre = "";
@@ -781,12 +621,6 @@ function displayGraph(id) {
 function removeTd(id) {
     d3.select("#frame_" + id).remove();
     removeInArray(openedGraph1, id);
-}
-
-function officialAQIus(data) {
-    const P1 = aqius(data.PM10, 'PM10');
-    const P2 = aqius(data.PM25, 'PM25');
-    return (P1 >= P2) ? {"AQI": P1, "origin": "PM10"} : {"AQI": P2, "origin": "PM2.5"};
 }
 
 function removeInArray(array) {
